Create some menu-Clean any code-Create modal form because frame have limit to Architect-

+++ booking/application.py
@@ -5,7 +5,7 @@ import tkinter as tk
 from tkinter import  *
 from tkinter import ttk, messagebox
 from datetime import datetime
-
+import traceback
 from .db.forms import RoomForm
 
 from . import db 
@@ -18,6 +18,8 @@ from sqlalchemy.ext.declarative import declarative_base
 class Application(tk.Tk):
     def __init__(self, *args ,**kwargs) :
         tk.Tk.__init__(self, *args, **kwargs)
+        
+        
         self._appconfig = AppConfig()
         self.settings = {}
         self.title('Hotel Booking ')
@@ -26,25 +28,15 @@ class Application(tk.Tk):
         engine = create_engine("postgresql://postgres:admin@localhost:5432/Booking", client_encoding="utf8",echo=True)
         self.Session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-        # At first run please create database Booking in to pgadmin4 then create database Schema by run this under command for create model 
-        # { init_db=  when manipulate models needs to active this schema to database server
-        # self.Base=db.models.Base
-        # Base=declarative_base(self.Base)
-        # self.Base.metadata.create_all(engine)
-        # }
-        
-        # self.init_db()
-        # from sqlalchemy import event
-        # from sqlalchemy.schema import CreateSchema
-
-        # event.listen(Base.metadata, 'before_create', CreateSchema('my_schema'))
-        
         self.callbacks = {
             # Menu option
             'file--quit': self.quit,
             'settings--preferences': self.open_preferences,
             'settings--preferences--update': self.update_preferences,
             'settings--preferences': self.open_preferences,
+            'settings--createschema': self.open_createschema,
+            'help--help':self.open_help,
+            'help--about':self.open_about,
             
             # Menu calls
             'open_room_form':self.open_room_form,
@@ -67,7 +59,10 @@ class Application(tk.Tk):
             'on_delete_reserve': self.on_delete_reserve,
             'on_clear_reserve': self.on_clear_reserve,
             'on_calc_reserve': self.on_calc_reserve,
+            'on_refresh_usertype': self.on_refresh_usertype,
+            
             'on_refresh_reserve_list': self.on_refresh_reserve_list,
+            'on_refresh_roomnumber_personid': self.on_refresh_roomnumber_personid,
             
             
             # User page functionality
@@ -135,6 +130,7 @@ class Application(tk.Tk):
 
         # Instance of Object- (View)
         self.app_form_window = None
+        
         self.room_form = None
         self.search_room_form = None
         self.reserve_form = None
@@ -143,9 +139,13 @@ class Application(tk.Tk):
         self.report_form = None
 
         self.room_view = None # 
+        self.room_view_modal = None
         
         self.preferences_form_window = None
         self.preferences_form = None
+        
+        self.search_form_window = None
+        
         
     @contextmanager
     def session_scope(self):
@@ -159,7 +159,13 @@ class Application(tk.Tk):
             raise
         finally:
             session.close()
-
+ 
+    def init_db(self):
+        self.Base=db.models.Base
+        Base=declarative_base(self.Base)
+        self.Base.metadata.create_all(engine)
+        return
+    
     def open_room_form(self):
         data={}
         if self.room_form is None:
@@ -192,14 +198,17 @@ class Application(tk.Tk):
             return   
    
     def on_delete_room(self):
-        data=self.room_form.get()
-        id = data['id']
-        try:
-            with self.session_scope() as session:
-                db.forms.RoomForm(session).delete_room(id)
-                self.room_form.delete_row()
-        except Exception as e:
-            messagebox.showinfo(title='Information',message='Not deleted selected row.')   
+        answer = messagebox.askyesno(title='Warning',
+                                     message='Are you sure delete this room information?\n Because by delete room information any reservation information for this room become [lost!!!]') 
+        if answer: 
+            data=self.room_form.get()
+            id = data['id']
+            try:
+                with self.session_scope() as session:
+                    db.forms.RoomForm(session).delete_room(id)
+                    self.room_form.delete_row()
+            except Exception as e:
+                messagebox.showinfo(title='Information',message='Not deleted selected row.')   
              
     def on_update_room(self):
         # switch state
@@ -248,6 +257,8 @@ class Application(tk.Tk):
             self.reserve_form.fetch_data(data)
             room_list = db.queries.qry_roomnumber_showall(session)
             self.reserve_form.list_room_number(room_list)
+            price_id_list = db.queries.qry_priceid_list(session)
+            self.reserve_form.list_price_id(price_id_list)
             
             self.reserve_form.grid(row=0, column=0, sticky='NSEW')
         else: 
@@ -255,26 +266,23 @@ class Application(tk.Tk):
             
     # == Reserve callback Contorls {     
     def on_add_reserve(self):
-        self.reserve_form.start_time()
-        self.reserve_form.end_time()  
-        self.start_datetime =self.reserve_form.datetime_start()
-        self.end_datetime =self.reserve_form.datetime_end()
+        self.reserve_form.start_time()#set variable spinbox time 
+        self.reserve_form.end_time()#set variable spinbox time   
+        self.start_datetime = self.reserve_form.datetime_start()# Gather date and time from widget => str(datetime)
+        self.end_datetime = self.reserve_form.datetime_end()# Gather date and time from widget => str(datetime)
         
         data = self.reserve_form.get()
         if data['roomnumber'] and data['personid'] and data['startdatetime'] and data['enddatetime'] and data['pricesum']:
-            
+            # convert str(datetime) to obj(datetime)
             data['startdatetime']=  datetime.strptime(self.start_datetime,'%m/%d/%y %H:%M:%S')
             data['enddatetime']=  datetime.strptime(self.end_datetime,'%m/%d/%y %H:%M:%S')
-            # print( 'start==',data['startdatetime'])
-            # print( 'end==',data['enddatetime'])
             try:
                 with self.session_scope() as session:
                     room_id = db.queries.qry_find_roomid_from_listroom(data['roomnumber'],session )#Show Form with combobox roomnumber and insert by id room
-                    idd=db.forms.ReserveForm(session).add_reserve(room_id, data)
-                    # print('== idd ==', idd)
+                    idd = db.forms.ReserveForm(session).add_reserve(room_id, data)
                     self.reserve_form.set(room_id)
-                    data = db.queries.qry_reserve_part_show(idd,session)
-                    self.reserve_form.add_row(idd,data)
+                    # self.reserve_form.add_row(idd,data)
+                    self.on_refresh_reserve()
                 messagebox.showinfo('Information','     Record saved      ')
             except NameError:
                 messagebox.showinfo('Warning','Something went wrong')
@@ -283,14 +291,17 @@ class Application(tk.Tk):
                 return    
     
     def on_delete_reserve(self):
-        data=self.reserve_form.get()
-        id= data['id']# need it check data[id] by  if id:
-        try:
-            with self.session_scope() as session:
-                db.forms.ReserveForm(session).delete_reserve(id)
-                self.reserve_form.delete_row()
-        except Exception as e:
-            messagebox.showinfo(title='Information',message='Not deleted selected row.')  
+        answer = messagebox.askyesno(title='Warning',
+                                     message='Are you sure delete this reserveation information?\n Because if delete this reservation information for that room state is => [Free!!!]') 
+        if answer: 
+            data=self.reserve_form.get()
+            id= data['id']# need it check data[id] by  if id:
+            try:
+                with self.session_scope() as session:
+                    db.forms.ReserveForm(session).delete_reserve(id)
+                    self.reserve_form.delete_row()
+            except Exception as e:
+                messagebox.showinfo(title='Information',message='Not deleted selected row.')  
             
     def on_update_reserve(self):
         pass
@@ -306,10 +317,12 @@ class Application(tk.Tk):
                 # find_id_price find id and price from room number and return by dict data keys=id and values=price
                 price=int(dicts['price'])
                 room_id=dicts['id']
+                print(price, room_id)
                 self.reserve_form.sets(price,room_id)
             if room_id and data['startdatetime'] and data['enddatetime']:
                 with self.session_scope() as session:
                     self.reserve_form.show_sumprice(price)
+                    # self.reserve_form.check_reserve_valid(room_id)
             else: 
                 messagebox.showinfo('Information','please Insert a Room Id,then start and end Date you have for Reservation.')
         except Exception as e:
@@ -340,6 +353,7 @@ class Application(tk.Tk):
     
     # { == User callback Contorls  User is Person table
     def on_add_user(self):
+        
         data = self.user_form.get()
         if data['usertype'] and data['name'] and data['family'] and data['email'] and data['tel'] and data['address']:
             try:
@@ -359,14 +373,17 @@ class Application(tk.Tk):
         return next((k for k, v in input_dict.items() if v == value), None)
     
     def on_delete_user(self):
-        data = self.user_form.get()
-        id= data['id']
-        try:
-            with self.session_scope() as session:
-                db.forms.UserForm(session).delete_user(id)
-                self.user_form.delete_row()
-        except Exception as e:
-            messagebox.showinfo(title='Information',message='Not deleted selected row.') 
+        answer = messagebox.askyesno(title='Warning',
+                                     message='Are you sure delete this user information?\n Because if delete this user information any reserve for this user state is => [Lost!!!]') 
+        if answer: 
+            data = self.user_form.get()
+            id= data['id']
+            try:
+                with self.session_scope() as session:
+                    db.forms.UserForm(session).delete_user(id)
+                    self.user_form.delete_row()
+            except Exception as e:
+                messagebox.showinfo(title='Information',message='Not deleted selected row.') 
              
     def on_update_user(self):
         pass
@@ -393,7 +410,7 @@ class Application(tk.Tk):
             self.usertype_form.lift()
             return
             
-    # { == User callback Contorls
+    # { == Usertype callback Contorls
     def on_add_usertype(self):
         data = self.usertype_form.get()
         if data['title']:
@@ -421,7 +438,7 @@ class Application(tk.Tk):
     def on_clear_usertype(self):
         self.usertype_form.clear()
         return
-    # == User callback Contorls }        
+    # == Usertype callback Contorls }        
    
     def open_report_view(self):
         if self.room_view is None:
@@ -435,44 +452,71 @@ class Application(tk.Tk):
         else:
             self.room_view.lift()
     
-    def on_refresh_reserve_list(self):
-        with self.session_scope() as session:
-            self.room_view = gui.views.Room_View( 
-                        self.workspace_frame,
-                        db.queries.qry_room_view(session),
-                        self.callbacks
-                            )
+    def on_refresh_reserve_list(self):# use Modal
+        if self.room_view_modal is None:
+            with self.session_scope() as session:
+                data = db.queries.qry_room_view(session)
+                self.room_view_modal = gui.views.Room_View_Modal( 
+                            self.app_form_window,
+                            data,
+                            self.callbacks
+                    )
+                self.room_view_modal.load_records()
         return 
+    
+    def on_refresh_roomnumber_personid(self):
+        
+        with self.session_scope() as session:
+            room_list = db.queries.qry_roomnumber_showall(session)
+            self.reserve_form.list_room_number(room_list)
+            price_id_list = db.queries.qry_priceid_list(session)
+            self.reserve_form.list_price_id(price_id_list)
+        return
             
      #  == Search callback Contorls 
     def open_search_room_form(self):
+       
         data={}
         if self.search_room_form is None:
             with self.session_scope() as session:
                 self.search_room_form = gui.forms.SearchRoomForm(self.workspace_frame,data,self.callbacks)
-                self.search_room_form.fetch_data(data)
             self.search_room_form.grid(row=0, column=0, sticky='NSEW')
         else:
             self.search_room_form.lift()
             
     def on_search_room_data(self):
-        data= self.search_room_form.get()
         try:
+            data = self.search_room_form.get()
+            print( data['searchby'],data['searchtext'])
             if data['searchby'] and data['searchtext']:
                 with self.session_scope() as session:
-                    self.search_room_form = gui.forms.SearchRoomForm(self.workspace_frame, data, self.callbacks)
-                    data2 = db.forms.SearchRoomForm(self).search( data, session)
-                # print('Dict==search===',data2.items())
-                self.search_room_form.fetch_data(data2)
-                data2={}
+                    self.search_room_form = gui.forms.SearchRoomModal(self.workspace_frame, data, self.callbacks)
+                    dict = db.forms.SearchRoomForm().search( data, session)
+                    print('Dict==search===',dict)
+                    self.search_room_form.fetch_data(dict)
+                    dict={}
+                    data={}
             else: 
                 messagebox.showinfo(title='search',message='Please Select search by and insert text for search.')
         except Exception as e:
-            messagebox.showerror('Error','You have an error in main search. \n{e}'.format(e))
+            traceback.print_exc()
+            messagebox.showerror('Error','You have an error in main search. \n')
             return
         finally:
             print('finally all search function comlete.')
+    
+    def on_refresh_usertype(self):
+        with self.session_scope() as session:
+            self.usertype_dict = db.queries.qry_usertype_list(session)
+            self.user_form.list_usertype(self.usertype_dict)
+        return
             
+    def on_refresh_reserve(self):
+        with self.session_scope() as session:
+            data = db.queries.qry_reserve_showall(session)
+            self.reserve_form.fetch_data(data)
+        return
+    
     def open_preferences(self):
         if self.preferences_form_window is None or not self.preferences_form_window.winfo_exists():
             self.preferences_form_window = tk.Toplevel(self)
@@ -488,6 +532,22 @@ class Application(tk.Tk):
     def update_preferences(self):
         data = self.preferences_form.appearance_frame.get()
         self._appconfig.update_settings(data)
+    
+    def open_createschema(self):
+        try:
+            self.init_db()
+        except Exception as e:
+            messagebox.showwarning('Error ','Fail to Generate schema for database (Booking). ')
+        else:
+            messagebox.showinfo('Info','Successfully to generate database schema.')
+        return
+    
+    def open_help(self):
+        messagebox.showinfo('Info','For first Run This application, please create Booking database in postgresql server and \n from Settings menu click [Create Database Schema] ')
+        return
+    
+    def open_about(self):
+         messagebox.showinfo('About','Thank you for use this application. \n Email:m_gehangir2006@gmail.com')
     # --------end
